+++ features/step_definitions/admin_steps/member/verify_with_scra_steps.rb
@@ -9,3 +9,15 @@ end
 Given(/^I should be successfully verified as a "([^"]*)"$/) do |group|
   VerifyWithScra.new.verify_scra_applied(@user_email, affiliation: group)
 end
+
+Given(/^I submit the empty SCRA form for "([^"]*)" in the Admin tool$/) do |group|
+  VerifyWithScra.new.verify_via_scra(@user_email, affiliation: group, populate: false)
+end
+
+Given(/^I should see error messages on required fields for "([^"]*)" in the Admin tool$/) do |affiliation|
+  VerifyWithScra.new.verify_admin_scra_form_error_messages(affiliation: affiliation)
+end
+
+Given(/^I should see red highlighted errors for "([^"]*)" in the Admin tool$/) do |affiliation|
+  VerifyWithScra.new.verify_admin_scra_form_red_highlighted_errors(affiliation: affiliation)
+end