+++ features/support/app_pages/admin/member/verify_with_scra.rb
@@ -1,16 +1,16 @@
-require_relative '../../base_classes/base_idp.rb'
+require_relative '../../base_classes/error_messages.rb'
 
 class VerifyWithScra < IDmeBase
 
+  include IVABase
   include Capybara::DSL
-  include IDPBase
-  include DataMagic
+  include ErrorMessages
 
   def initialize
     super("#{FigNewton.admin.user_quick_search}")
   end
 
-  def verify_via_scra(username, affiliation:, populate: true, type: nil)
+  def verify_via_scra(user_email, affiliation:, populate: true, type: nil)
     case affiliation
     when "Service Member"   then data_set = :scra_veteran
     when "Veteran"          then data_set = :scra_veteran
@@ -22,10 +22,11 @@ class VerifyWithScra < IDmeBase
 
     data = data_for(data_set)
 
+    search_for_user(user_email)
+    select affiliation
+
     if populate
 
-      search_for_user(username)
-      select affiliation
       populate_form_fields(data: data)
       if["Military Spouse", "Military Family"].include?(affiliation)
         2.times {
@@ -35,19 +36,20 @@ class VerifyWithScra < IDmeBase
           fill_in "scra_request_service_date", :with => data.fetch("verification_service_date")
         }
       end
-      click_button("Submit")
     end
+
+    click_button("Submit")
   end
 
-  def verify_scra_applied(username, affiliation:)
+  def verify_scra_applied(user_email, affiliation:)
     sleep 1
     all('tr')[1].all('a')[0].click
     sleep 2
-    page.assert_text username
+    page.assert_text user_email
   end
 
-  def search_for_user(username)
-    populate_search_field(search_box: username)
+  def search_for_user(user_email)
+    populate_search_field(search_box: user_email)
     click_button("Search")
     click_link("capybara+")
     click_link("Verify with SCRA")
@@ -63,5 +65,44 @@ class VerifyWithScra < IDmeBase
     fill_in "scra_request_service_member_birth_date", :with => data.fetch("verification_service_member_birth_date")
     fill_in "scra_request_social", :with => data.fetch("verification_social")
     fill_in "scra_request_service_date", :with => data.fetch("verification_service_date")
+  end
+
+  def verify_admin_scra_form_error_messages(affiliation:)
+    form_field = %w[service_member_first_name service_member_last_name service_member_birth_date social service_date]
+    form_field_family = %w[first_name last_name birth_date]
+
+    fields = case affiliation
+             when "Service Member", "Veteran", "Retiree"    then form_field
+             when "Military Family", "Military Spouse"      then form_field.push(*form_field_family)
+             else fail("Affiliation not found")
+             end
+
+   form_error_messages_count.should == fields.count
+  end
+
+  def verify_admin_scra_form_red_highlighted_errors(affiliation:)
+    form_field = %w[service_member_first_name service_member_last_name service_member_birth_date social service_date]
+    form_field_family = %w[first_name last_name birth_date]
+
+    fields = case affiliation
+             when "Service Member", "Veteran", "Retiree"    then form_field
+             when "Military Family", "Military Spouse"      then form_field.push(*form_field_family)
+             else fail("Affiliation not found")
+             end
+
+    fields.each do |field|
+      %w(top right bottom left).each do |border_side|
+        red_highlighted_field = page.find("#scra_request_#{field}").native.css_value("border-#{border_side}-color")
+        red_highlighted_field.should == admin_red
+      end
+    end
+  end
+
+  def form_error_messages_count
+    all(".formError").count
+  end
+
+  def admin_red
+    "rgba(185, 74, 72, 1)"
   end
 end